some tinkering

+++ dirty_images.py
@@ -9,35 +9,41 @@ sdir = '/data/sandrews/ALMA_regridding/storage/'
 name = 'ALMA-WSU_native'
 binn = '.bin8x'
 
+# example for demo WSU post-averaging setup
+name = 'ALMA-WSU_native'
+binn = '_postavg.bin8x'
+
 # example for demo BLC setup
 #name = 'ALMA-BLC_122kHz'
 #binn = ''
 
+img_truth = False
 nm_f = ['05']
 imeth = ['nearest', 'linear', 'cubic', 'fftshift']
 
 
 for ig in range(len(nm_f)):
 
-    ### regridded TRUTH
-    prefix = 'TRUTH.'+name+binn+'.REGRID'+nm_f[ig]
+    if img_truth:
+        ### regridded TRUTH
+        prefix = 'TRUTH.'+name+binn+'.REGRID'+nm_f[ig]
 
-    # prepare / cleanup
-    ext = ['image', 'model', 'pb', 'psf', 'residual', 'sumwt']
-    for i in range(len(ext)):
-        if os.path.exists(sdir+'images/'+prefix+'.cube.'+ext[i]):
-            os.system('rm -rf '+sdir+'images/'+prefix+'.cube.'+ext[i])
+        # prepare / cleanup
+        ext = ['image', 'model', 'pb', 'psf', 'residual', 'sumwt']
+        for i in range(len(ext)):
+            if os.path.exists(sdir+'images/'+prefix+'.cube.'+ext[i]):
+                os.system('rm -rf '+sdir+'images/'+prefix+'.cube.'+ext[i])
 
-    # dirty cube
-    tclean(vis=sdir+prefix+'.ms', 
-           imagename=sdir+'images/'+prefix+'.cube',
-           specmode='cubedata', imsize=512, cell='0.02arcsec', niter=0,
-           restoringbeam='common')
+        # dirty cube
+        tclean(vis=sdir+prefix+'.ms', 
+               imagename=sdir+'images/'+prefix+'.cube',
+               specmode='cubedata', imsize=512, cell='0.02arcsec', niter=0,
+               restoringbeam='common')
 
-    # make a fits cube
-    exportfits(sdir+'images/'+prefix+'.cube.image', 
-               fitsimage=sdir+'images/'+prefix+'.cube.image.fits',
-               overwrite=True)
+        # make a fits cube
+        exportfits(sdir+'images/'+prefix+'.cube.image', 
+                   fitsimage=sdir+'images/'+prefix+'.cube.image.fits',
+                   overwrite=True)
 
 
     # iterate over interpolation methods
